Extends CSUI login through all apps

+++ lab_5/views.py
@@ -3,20 +3,24 @@ from django.contrib import messages
 from django.http import HttpResponseRedirect
 from .forms import Todo_Form
 from .models import Todo
+from praktikum.custom_auth import check_login
 
 # Create your views here.
 response = {}
-def index(request):    
+def index(request):
 	response['author'] = "Ichlasul Affan" #TODO Implement yourname
 	todo = Todo.objects.all()
 	response['todo'] = todo
 	html = 'lab_5/lab_5.html'
+	html = check_login(request, html, response)
 	response['todo_form'] = Todo_Form
 	return render(request, html, response)
 
 def add_todo(request):
 	form = Todo_Form(request.POST or None)
-	if(request.method == 'POST' and form.is_valid()):
+	dummy_response = check_login(request, False, response)
+	if dummy_response: return render(request, dummy_response, response)
+	elif request.method == 'POST' and form.is_valid():
 		response['title'] = request.POST['title']
 		response['description'] = request.POST['description']
 		todo = Todo(title=response['title'],description=response['description'])
@@ -29,10 +33,13 @@ def add_todo(request):
 		return HttpResponseRedirect('/lab-5/')
 
 def delete_todo(request):
-	try:
-		object = Todo.objects.get(id=request.GET.get('id',''))
-		object.delete()
-		messages.add_message(request, messages.SUCCESS, "Todo has been deleted.", extra_tags="delete")
-	except:
-		messages.add_message(request, messages.ERROR, "The todo does not exist.", extra_tags="delete")
-	finally: return HttpResponseRedirect('/lab-5/#my-list')
+	dummy_response = check_login(request, False, response)
+	if dummy_response: return render(request, dummy_response, response)
+	else:
+		try:
+			object = Todo.objects.get(id=request.GET.get('id',''))
+			object.delete()
+			messages.add_message(request, messages.SUCCESS, "Todo has been deleted.", extra_tags="delete")
+		except:
+			messages.add_message(request, messages.ERROR, "The todo does not exist.", extra_tags="delete")
+		finally: return HttpResponseRedirect('/lab-5/#my-list')
